requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
djangorestframework==3.14.0
idna==3.6
pytz==2023.3.post1
razorpay==1.4.1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
